hci-collector: replaced error.h that is missing on OpenWrt
Unfortunatelly, the error.h header file is missing on the OpenWrt 18.06.
There was a similar issue elsewhere
(https://github.com/hzeller/gmrender-resurrect/issues/107)
and the solution there was to replace the function with different one.

This patch replaces perror() as well; errno is printed instead of the
"system" error message.